Move the IContactAndReservationSettings to adhere to the Dependency Inversion Principle
